Integrate LibreSSL + portability
As the mrbgem depends on a SSL library such as OpenSSL or LibreSSL, it was
tedious to have to deal with that manually. This change automates the building
of LibreSSL when building the mrbgem.

The change deals also with portability to the targets linux, osx, win and
architecture 32 and 64 bits.